feat(mcp-oauth): scaffold - provider dep, OAUTH_KV bindings, post-auth pipeline extraction
BL-033 Slice 2, Phase 1 of 6. Zero behavior change - full suite green
(1718/1718) before any OAuth code lands.

- @cloudflare/workers-oauth-provider pinned exact at 0.8.2. Installed
  package verified against the planning research: GrantType has NO
  client_credentials (M2M stays our own token branch per plan D6),
  apiRoute accepts string[], DCR off by omitting
  clientRegistrationEndpoint, CIMD via clientIdMetadataDocumentEnabled,
  PRM served natively.
- wrangler.toml: OAUTH_KV namespaces bound per env (created via wrangler
  2026-07-24; staging id 580b8f1f..., production 13c9e9f5...), plus the
  global_fetch_strictly_public compat flag - SSRF defense for CIMD's
  fetch of client-supplied metadata URLs, not just a no-impact toggle.
- worker.ts fetch handler: the post-authentication block (Sentry tag ->
  rate limit -> /radar/snapshot -> MCP dispatch -> request log -> RL+CORS
  headers) moves verbatim to pipeline/handle-authenticated.ts keyed on
  the AuthSuccess contract, so the upcoming OAuth and M2M auth paths
  converge on byte-identical downstream behavior. Extracting to a
  separate module (not worker.ts) pre-empts the worker->oauth->worker
  import cycle.
- Env: OAUTH_KV + OAUTH_M2M_SIGNING_KEY declared (staging signing key
  already bound via wrangler secret put, stdin pipe).
