+++ mcp-server/package.json
@@ -29,6 +29,7 @@
   },
   "dependencies": {
     "@cfworker/json-schema": "^4.1.1",
+    "@cloudflare/workers-oauth-provider": "0.8.2",
     "@modelcontextprotocol/sdk": "^1.29.0",
     "@sentry/cloudflare": "^10.65.0",
     "@upstash/ratelimit": "^2.0.5",
